fix lint bug
